refactor: Improve date localization and formatting
