+++ clients/tablet/core/domain/src/commonMain/kotlin/band/effective/office/tablet/core/domain/util/LocalDateTimeFormatter.kt
@@ -30,5 +30,3 @@ fun LocalTime.toFormattedString(pattern: String): String {
     }
     return formatter.format(this)
 }
-
-expect fun LocalDateTime.toLocalisedString(pattern: String): String
