+++ clients/tablet/core/ui/build.gradle.kts
@@ -10,7 +10,7 @@ kotlin {
             implementation(libs.decompose.compose.jetbrains)
             implementation(libs.essenty.lifecycle)
             implementation(libs.essenty.state.keeper)
-            implementation(project(":clients:tablet:core:domain"))
+            api(libs.kotlinx.datetime)
         }
     }
 }
