+++ clients/tablet/core/ui/src/androidMain/kotlin/band/effective/office/tablet/core/ui/util/DateFormatter.android.kt
@@ -1,4 +1,4 @@
-package band.effective.office.tablet.core.domain.util
+package band.effective.office.tablet.core.ui.utils
 
 import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.toJavaLocalDateTime
