+++ clients/tablet/core/ui/src/commonMain/kotlin/band/effective/office/tablet/core/ui/utils/DateFormatter.kt
@@ -1,0 +1,6 @@
+package band.effective.office.tablet.core.ui.utils
+
+import kotlinx.datetime.LocalDateTime
+
+@OptIn(kotlinx.datetime.format.FormatStringsInDatetimeFormats::class)
+expect fun LocalDateTime.toLocalisedString(pattern: String): String
