+++ clients/tablet/core/ui/src/iosMain/kotlin/band/effective/office/tablet/core/ui/utils/DateFormatter.ios.kt
@@ -1,4 +1,4 @@
-package band.effective.office.tablet.core.domain.util
+package band.effective.office.tablet.core.ui.utils
 
 import kotlinx.datetime.LocalDateTime
 import platform.Foundation.NSDate
